remove duplication in feat gen script
